requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
bandit==1.6.2
certifi==2019.11.28
chardet==3.0.4
chromedriver-py==83.0.4103.39
discord==1.0.1
discord.py==1.3.1
gitdb==4.0.5
GitPython==3.1.7
idna==2.8
lxml==4.5.0
multidict==4.7.4
pbr==5.4.5
python-dotenv==0.11.0
PyYAML==5.3.1
requests==2.22.0
selenium==3.141.0
six==1.15.0
smmap==3.0.4
stevedore==3.2.0
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
